Pushing register map parsing code into Registers.scala so that Apb2CSTrgt.scal and GenCSTrgt.scala are (mainly) Chisel. Refactoring Apb2CSTrgt and GenCSTrgt to push all logic into GenCSTrgt (e.g. register aliasing)
